New function/Fix bug
Now we have 12 cards left uncompleted and we can distribute cards at the beginning of each round.
now we can't choose ourselves when choosing next player, and our cards will be reduce correctly after we reveal them.
We still need to implement virtual players and improve the effects of the cards